Split out slang-serialize-misc-type-info.h
