+++ source/slang/slang-serialize-ast.cpp
@@ -4,26 +4,19 @@
 #include "slang-ast-generated.h"
 #include "slang-ast-generated-macro.h"
 
-#include "slang-compiler.h"
-#include "slang-type-layout.h"
-
 #include "slang-ast-dump.h"
-#include "slang-mangle.h"
 
 #include "slang-ast-support-types.h"
 
+// Needed for ModuleSerialFilter
+// Needed for 'findModuleForDecl'
 #include "slang-legalize-types.h"
-
-#include "slang-serialize.h"
-
-#include "slang-source-loc.h"
+#include "slang-mangle.h"
 
 #include "slang-serialize-type-info.h"
-
-#include "../core/slang-byte-encode-util.h"
+#include "slang-serialize-misc-type-info.h"
 
 namespace Slang {
-
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! ModuleSerialFilter  !!!!!!!!!!!!!!!!!!!!!!!!
 
@@ -77,181 +70,6 @@ SerialIndex ModuleSerialFilter::writePointer(SerialWriter* writer, const NodeBas
     // For now for everything else just write it
     return writer->writeObject(ptr);
 }
-
-// !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! Serial <-> Native conversion  !!!!!!!!!!!!!!!!!!!!!!!!
-
-// Collection of other types
-
-// SamplerStateFlavor
-
-template <>
-struct SerialTypeInfo<SamplerStateFlavor> : public SerialConvertTypeInfo<SamplerStateFlavor, uint8_t> {};
-
-// TextureFlavor
-
-template <>
-struct SerialTypeInfo<TextureFlavor>
-{
-    typedef TextureFlavor NativeType;
-    typedef uint16_t SerialType;
-    enum { SerialAlignment = sizeof(SerialType) };
-
-    static void toSerial(SerialWriter* writer, const void* native, void* serial) { SLANG_UNUSED(writer); *(SerialType*)serial = ((const NativeType*)native)->flavor; }
-    static void toNative(SerialReader* reader, const void* serial, void* native) { SLANG_UNUSED(reader); ((NativeType*)native)->flavor = *(const SerialType*)serial; }
-};
-
-// ImageFormat
-template <>
-struct SerialTypeInfo<ImageFormat> : public SerialConvertTypeInfo<ImageFormat, uint8_t> {};
-
-// Stage
-template <>
-struct SerialTypeInfo<Stage> : public SerialConvertTypeInfo<Stage, uint8_t> {};
-
-// TokenType
-template <>
-struct SerialTypeInfo<TokenType> : public SerialConvertTypeInfo<TokenType, uint8_t> {};
-
-// BaseType
-template <>
-struct SerialTypeInfo<BaseType> : public SerialConvertTypeInfo<BaseType, uint8_t> {};
-
-// SemanticVersion
-template <>
-struct SerialTypeInfo<SemanticVersion> : public SerialIdentityTypeInfo<SemanticVersion> {};
-
-// SourceLoc
-
-// Make the type exposed, so we can look for it if we want to remap.
-template <>
-struct SerialTypeInfo<SourceLoc>
-{
-    typedef SourceLoc NativeType;
-    typedef SerialSourceLoc SerialType;
-    enum { SerialAlignment = SLANG_ALIGN_OF(SerialSourceLoc) };
-
-    static void toSerial(SerialWriter* writer, const void* inNative, void* outSerial)
-    {
-        DebugSerialWriter* debugWriter = writer->getExtraObjects().get<DebugSerialWriter>();
-        *(SerialType*)outSerial = debugWriter ? debugWriter->addSourceLoc(*(const NativeType*)inNative) : SerialType(0);
-    }
-    static void toNative(SerialReader* reader, const void* inSerial, void* outNative)
-    {
-        DebugSerialReader* debugReader = reader->getExtraObjects().get<DebugSerialReader>();
-        *(NativeType*)outNative = debugReader ? debugReader->getSourceLoc(*(const SerialType*)inSerial) : NativeType::fromRaw(0);
-    }
-};
-
-// Token
-template <>
-struct SerialTypeInfo<Token>
-{
-    typedef Token NativeType;
-    struct SerialType
-    {
-        SerialTypeInfo<BaseType>::SerialType type;
-        SerialTypeInfo<SourceLoc>::SerialType loc;
-        SerialIndex name;
-    };
-    enum { SerialAlignment = SLANG_ALIGN_OF(SerialType) };
-
-    static void toSerial(SerialWriter* writer, const void* native, void* serial)
-    {
-        auto& src = *(const NativeType*)native;
-        auto& dst = *(SerialType*)serial;
-
-        SerialTypeInfo<TokenType>::toSerial(writer, &src.type, &dst.type);
-        SerialTypeInfo<SourceLoc>::toSerial(writer, &src.loc, &dst.loc);
-
-        if (src.flags & TokenFlag::Name)
-        {
-            dst.name = writer->addName(src.getName());
-        }
-        else
-        {
-            dst.name = writer->addString(src.getContent());
-        }
-    }
-    static void toNative(SerialReader* reader, const void* serial, void* native)
-    {
-        auto& src = *(const SerialType*)serial;
-        auto& dst = *(NativeType*)native;
-
-        dst.flags = 0;
-        dst.charsNameUnion.chars = nullptr;
-
-        SerialTypeInfo<TokenType>::toNative(reader, &src.type, &dst.type);
-        SerialTypeInfo<SourceLoc>::toNative(reader, &src.loc, &dst.loc);
-
-        // At the other end all token content will appear as Names.
-        if (src.name != SerialIndex(0))
-        {
-            dst.charsNameUnion.name = reader->getName(src.name);
-            dst.flags |= TokenFlag::Name;
-        }
-    }
-};
-
-// NameLoc
-template <>
-struct SerialTypeInfo<NameLoc>
-{
-    typedef NameLoc NativeType;
-    struct SerialType
-    {
-        SerialTypeInfo<SourceLoc>::SerialType loc;
-        SerialIndex name;
-    };
-    enum { SerialAlignment = SLANG_ALIGN_OF(SerialType) };
-
-    static void toSerial(SerialWriter* writer, const void* native, void* serial)
-    {
-        auto& src = *(const NativeType*)native;
-        auto& dst = *(SerialType*)serial;
-
-        dst.name = writer->addName(src.name);
-        SerialTypeInfo<SourceLoc>::toSerial(writer, &src.loc, &dst.loc);
-    }
-    static void toNative(SerialReader* reader, const void* serial, void* native)
-    {
-        auto& src = *(const SerialType*)serial;
-        auto& dst = *(NativeType*)native;
-
-        dst.name = reader->getName(src.name);
-        SerialTypeInfo<SourceLoc>::toNative(reader, &src.loc, &dst.loc);
-    }
-};
-
-// DiagnosticInfo
-template <>
-struct SerialTypeInfo<const DiagnosticInfo*>
-{
-    typedef const DiagnosticInfo* NativeType;
-    typedef SerialIndex SerialType;
-
-    enum { SerialAlignment = SLANG_ALIGN_OF(SerialType) };
-
-    static void toSerial(SerialWriter* writer, const void* native, void* serial)
-    {
-        auto& src = *(const NativeType*)native;
-        auto& dst = *(SerialType*)serial;
-        dst = src ? writer->addString(UnownedStringSlice(src->name)) : SerialIndex(0);
-    }
-    static void toNative(SerialReader* reader, const void* serial, void* native)
-    {
-        auto& src = *(const SerialType*)serial;
-        auto& dst = *(NativeType*)native;
-
-        if (src == SerialIndex(0))
-        {
-            dst = nullptr;
-        }
-        else
-        {
-            dst = findDiagnosticByName(reader->getStringSlice(src));
-        }
-    }
-};
 
 /* !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! AST types !!!!!!!!!!!!!!!!!!!!!!!!!!!!! */
