+++ source/slang/slang.vcxproj
@@ -288,6 +288,7 @@
     <ClInclude Include="slang-serialize-debug.h" />
     <ClInclude Include="slang-serialize-ir-types.h" />
     <ClInclude Include="slang-serialize-ir.h" />
+    <ClInclude Include="slang-serialize-misc-type-info.h" />
     <ClInclude Include="slang-serialize-reflection.h" />
     <ClInclude Include="slang-serialize-type-info.h" />
     <ClInclude Include="slang-serialize-types.h" />
